backend/requirements.txt
fastapi==0.115.2
aiosmtplib==3.0.2
uvicorn==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic_core==2.23.4
starlette==0.40.0
openai==1.52.0
certifi==2024.8.30
idna==3.10
loguru==0.7.2
python-dotenv==1.0.1
setuptools==75.1.0
supabase==2.9.1
storage3==0.8.2
postgrest==0.17.2
gotrue==2.9.3
realtime==2.0.6
